pizza-asif/pyproject.toml
[tool.poetry]
name = "pizza-asif"
version = "0.1.0"
description = ""
authors = ["asif-langrah <ak.development.pk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
uvicorn = "^0.30.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
